+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog (Pypi package)
 
+### Fixed
+
+- Google credentials: better error management in case `secret_key` or `secret_key_id` is `None`
+
 ## [3.23.25] 2024-01-17
 
 ### Fixed